% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_web_mapping.R
\name{addDiffLegend}
\alias{addDiffLegend}
\title{Add a continuous legend for a difference raster (with controlled domain)}
\usage{
addDiffLegend(
  map,
  diff_raster,
  title,
  layerId,
  palette = NULL,
  diff_range = NULL,
  diff_quantiles = c(0, 1)
)
}
\description{
Add a continuous legend for a difference raster (with controlled domain)
}
\keyword{internal}
