% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_web_mapping.R
\name{addDiffRasterLayer}
\alias{addDiffRasterLayer}
\title{Add a continuous difference raster layer (with controlled domain)}
\usage{
addDiffRasterLayer(
  map,
  diff_raster,
  group,
  palette = NULL,
  diff_range = NULL,
  diff_quantiles = c(0, 1)
)
}
\description{
Add a continuous difference raster layer (with controlled domain)
}
\keyword{internal}
