% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_web_mapping.R
\name{addLegendLayer}
\alias{addLegendLayer}
\title{Add a legend for a continuous raster layer}
\usage{
addLegendLayer(map, raster, title, layerId, color_palette)

addLegendLayer(map, raster, title, layerId, color_palette)
}
\description{
Add a legend for a continuous raster layer

Add a legend for a continuous raster layer
}
\keyword{internal}
