% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_web_mapping.R
\name{addRasterLayer}
\alias{addRasterLayer}
\title{Add a continuous raster layer to a leaflet map}
\usage{
addRasterLayer(map, raster, name, color_palette)
}
\description{
Add a continuous raster layer to a leaflet map
}
\keyword{internal}
