% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_raster_polygon.R
\name{cfcore_validate_breaks}
\alias{cfcore_validate_breaks}
\title{Validate breaks for binning}
\usage{
cfcore_validate_breaks(breaks)
}
\arguments{
\item{breaks}{Numeric vector of strictly increasing finite breakpoints.}
}
\description{
Validate breaks for binning
}
\keyword{internal}
