% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudFlux.R
\docType{class}
\name{cloudFlux}
\alias{cloudFlux}
\alias{cloudFlux-class}
\title{cloudFlux reference class}
\description{
A mutable workflow container that orchestrates two-epoch point cloud change
processing in CFCore: mask generation, denoising, DTM/nDSM derivation, raster
alignment, and continuous difference computation.
}
\keyword{internal}
