% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{create_directories}
\alias{create_directories}
\title{Create a directory (recursively) if it does not exist}
\usage{
create_directories(in_dir, verbose = FALSE)
}
\arguments{
\item{in_dir}{Character scalar. Directory path to create.}

\item{verbose}{Logical. If \code{TRUE}, prints a message.}
}
\value{
Invisibly returns \code{in_dir}.
}
\description{
Create a directory (recursively) if it does not exist
}
\keyword{internal}
