% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{delete_all_files}
\alias{delete_all_files}
\title{Delete files in a directory}
\usage{
delete_all_files(dir, pattern = NULL, recursive = FALSE, dry_run = FALSE)
}
\arguments{
\item{dir}{Character scalar. Directory to delete files from.}

\item{pattern}{Optional regex pattern to filter files. Defaults to \code{NULL} (all files).}

\item{recursive}{Logical. If \code{TRUE}, includes files in subdirectories.}

\item{dry_run}{Logical. If \code{TRUE}, returns the files that would be deleted without deleting.}
}
\value{
Character vector of files deleted (or that would be deleted if \code{dry_run=TRUE}).
}
\description{
Deletes files in a directory. Use with care.
}
\keyword{internal}
