% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_raster_polygon.R
\name{diff_dtm}
\alias{diff_dtm}
\title{Continuous DTM difference (later - earlier)}
\usage{
diff_dtm(earlier, later)
}
\arguments{
\item{earlier}{\code{terra::SpatRaster} (earlier epoch).}

\item{later}{\code{terra::SpatRaster} (later epoch).}
}
\value{
\code{terra::SpatRaster} of continuous change values.
}
\description{
Continuous DTM difference (later - earlier)
}
\keyword{internal}
