% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_web_mapping.R
\name{displayIndex}
\alias{displayIndex}
\title{Display index polygons for loaded spatial containers}
\usage{
displayIndex(index, epsg = 4326, measure_colors = c("#3D535D", "#7D4479"))
}
\description{
Display index polygons for loaded spatial containers
}
\keyword{internal}
