% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_web_mapping.R
\name{displayMap}
\alias{displayMap}
\title{Display a leaflet map with rasters and optional continuous change layers}
\usage{
displayMap(
  dtm1,
  ndsm1,
  dtm2,
  ndsm2,
  dtm_diff,
  ndsm_diff,
  area_mask,
  epsg = 4326,
  diff_palette = NULL,
  diff_range = NULL,
  diff_quantiles = c(0, 1),
  diff_breaks = NULL,
  diff_break_colors = NULL,
  measure_colors = c("#3D535D", "#7D4479")
)
}
\description{
Display a leaflet map with rasters and optional continuous change layers
}
\keyword{internal}
