% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_raster_polygon.R
\name{is_empty}
\alias{is_empty}
\title{Check whether a SpatRaster is empty (all NA)}
\usage{
is_empty(raster)
}
\arguments{
\item{raster}{\code{terra::SpatRaster}.}
}
\value{
\code{TRUE} if all raster values are NA, otherwise \code{FALSE}.
}
\description{
Check whether a SpatRaster is empty (all NA)
}
\keyword{internal}
