% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pointcloud.R
\name{noise_filter}
\alias{noise_filter}
\title{Denoise a LAS point cloud using iterative SOR noise classification}
\usage{
noise_filter(laz, k_sor1 = 5, m_sor1 = 3, k_sor2 = 20, m_sor2 = 5)
}
\arguments{
\item{laz}{A \code{lidR::LAS} object.}

\item{k_sor1, m_sor1}{SOR parameters for first pass (\code{lidR::sor(k=..., m=...)}).}

\item{k_sor2, m_sor2}{SOR parameters for second pass.}
}
\value{
A \code{lidR::LAS} object with noise filtered.
}
\description{
Runs iterative SOR noise classification and removes points classified as noise
(Classification == 18).
}
\keyword{internal}
