% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_raster_polygon.R
\name{plot_diff_binned}
\alias{plot_diff_binned}
\title{Plot binned statistics for a continuous difference raster}
\usage{
plot_diff_binned(
  diff_raster,
  breaks,
  palette = NULL,
  title = "Difference Raster Statistics",
  xlab = "Change bin"
)
}
\arguments{
\item{diff_raster}{\code{terra::SpatRaster} continuous change.}

\item{breaks}{Numeric vector of breakpoints.}

\item{palette}{Optional vector of colors (length = number of bins).}

\item{title}{Plot title.}

\item{xlab}{X-axis label.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot binned statistics for a continuous difference raster
}
\keyword{internal}
