% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_raster_polygon.R
\name{plot_diff_hist}
\alias{plot_diff_hist}
\title{Plot a histogram for a continuous difference raster}
\usage{
plot_diff_hist(
  diff_raster,
  bins = 60,
  title = "Difference Histogram",
  xlab = "Change"
)
}
\arguments{
\item{diff_raster}{\code{terra::SpatRaster} continuous change.}

\item{bins}{Integer number of histogram bins.}

\item{title}{Plot title.}

\item{xlab}{X-axis label.}
}
\value{
A \code{ggplot2} object.
}
\description{
This is the pure continuous view (no bins by semantic breaks).
}
\keyword{internal}
