% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_web_mapping.R
\name{processRasterMask}
\alias{processRasterMask}
\title{Project and optionally mask a raster for web display}
\usage{
processRasterMask(raster, mask, epsg = 4326)
}
\description{
Project and optionally mask a raster for web display
}
\keyword{internal}
