% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-TTP.R
\name{TTP_23}
\alias{TTP_23}
\title{Downsampled LiDAR point cloud – Tommy Thompson Park (2023)}
\format{
A \code{\link[lidR:LAS-class]{lidR::LAS}} object.
}
\source{
Internal example data derived from City of Toronto LiDAR.
}
\description{
A downsampled airborne LiDAR point cloud representing Tommy Thompson Park,
collected in 2023. This dataset pairs with \link{TTP_15} to demonstrate
multi-temporal LiDAR change workflows in CloudFluxCore.
}
\seealso{
\link{TTP_15}
}
\keyword{datasets}
