% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_web_mapping.R
\name{addDiffLegend}
\alias{addDiffLegend}
\title{Add a continuous legend for a difference raster (with controlled domain)}
\usage{
addDiffLegend(
  map,
  diff_raster,
  title,
  layerId,
  palette = NULL,
  diff_range = NULL,
  diff_quantiles = c(0.02, 0.98)
)
}
\arguments{
\item{map}{Leaflet map.}

\item{diff_raster}{\code{terra::SpatRaster} continuous difference raster.}

\item{title}{Legend title.}

\item{layerId}{Leaflet layer ID.}

\item{palette}{Palette name or vector accepted by \code{leaflet::colorNumeric()}.}

\item{diff_range}{Optional numeric length-2 c(min, max) for domain.}

\item{diff_quantiles}{Quantiles used for domain if \code{diff_range} is NULL.}
}
\value{
Leaflet map.
}
\description{
Add a continuous legend for a difference raster (with controlled domain)
}
\keyword{internal}
