% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_web_mapping.R
\name{addLegendLayer}
\alias{addLegendLayer}
\title{Add a legend for a continuous raster layer}
\usage{
addLegendLayer(map, raster, title, layerId, color_palette)
}
\arguments{
\item{map}{Leaflet map.}

\item{raster}{\code{terra::SpatRaster} or \code{NULL}.}

\item{title}{Legend title.}

\item{layerId}{Leaflet layer ID.}

\item{color_palette}{Palette name or vector accepted by \code{leaflet::colorNumeric()}.}
}
\value{
Leaflet map.
}
\description{
Add a legend for a continuous raster layer
}
\keyword{internal}
