% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_web_mapping.R
\name{addRasterLayer}
\alias{addRasterLayer}
\title{Add a continuous raster layer to a leaflet map}
\usage{
addRasterLayer(map, raster, name, color_palette)
}
\arguments{
\item{map}{Leaflet map.}

\item{raster}{\code{terra::SpatRaster} or \code{NULL}.}

\item{name}{Group name for leaflet layer control.}

\item{color_palette}{Palette name or vector accepted by \code{leaflet::colorNumeric()}.}
}
\value{
Leaflet map.
}
\description{
Add a continuous raster layer to a leaflet map
}
\keyword{internal}
