% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudFlux.R
\name{cloudFlux_new}
\alias{cloudFlux_new}
\title{Create a cloudFlux object}
\usage{
cloudFlux_new(
  source_path = character(0),
  target_path = character(0),
  source_las = NULL,
  target_las = NULL,
  epsg = 26917,
  resolution = 1,
  use_icp = FALSE,
  use_gpu = TRUE,
  voxel_size = 0.05,
  icp_method = "point-to-plane"
)
}
\description{
Create a cloudFlux object
}
